This is still WIP.

- Do not copy the install prefix before tarballing it (except when --rel, but
  who uses that?)
- Use zstd for (de)compression

For me this results in the following:

LLVM + deps (7.8GB):

Before:
- `buildcache create`: `5m 35s`
- buildcache size: 2.6GB
- `install`: 47s

After:
- `buildcache create`: `49s`
- buildcache size: 2.4GB
- `install`: 24s